requirements.txt
Django==3.1.5
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
reportlab==3.5.34
requests==2.22.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1

